remove deprecated proprety namespace collisions
require writing `new_resource.my_property` instead of just
`my_property` in provider code.

a surprise in this code was that methods directly on the
resource are all injected into the provider, and we never
deprecated that, because we didn't know about it.

that makes `action_class {}` helpers unnecessary, but causes
additional namespace pollution, but at least the methods declared
in the resource will lose to DSL methods (i.e. `def execute`
will not work).

but that also means that resources which are injected into the DSL
which overwrite resource methods in other cookbooks may break
their use in that cookbooks providers/action_classes.

we should probably deprecate that as well, but we didn't and
people probably found this behavior and used it.

i cleaned it up a bit and removed the methods on Chef::Resource
directly which polluted the provider class with all kinds of
bizzare crap methods (not_if and only_if were getting injected
into the provider, etc).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>