Pass pointer to LsaFreeMemory, not FFI::MemoryPointer
LsaFreeMemory takes a pointer to the struct in memory created by LsaEnumerateAccountRights.
We have to pass it the actual pointer, not the FFI::MemoryPointer object.

This fixes masked access violations which occasionally lead to a heap corruption which
presents itself as silent termination of chef-client/ruby with a return code of -1073740940
or STATUS_HEAP_CORRUPTION / 0xc0000374.

Fixes #6589

Signed-off-by: Bryan McLellan <btm@loftninjas.org>