Remove Chef 12-isms from the apt_repository resource
We set a lot of nil defaults and did coersions here that are no longer necessary. Since we did accept nil in a lot of these properties at one point we need to keep it so I wrote a note to future me about that. I also added some specs for the remaining coersion and the defaults.

Signed-off-by: Tim Smith <tsmith@chef.io>