Support installing removed windows features from source
This issue was reported on the cookbook after refactoring the resource. The issue is that we have a method to determine the features to install given a list of features we want to install. It wasn't taking into account that a removed feature can be installed if the user specifies a source. Now it properly takes that into account and should build the correct array of features to install. It also builds the install command in a cleaner way that avoids funky extra spaces at the end of the command.

Signed-off-by: Tim Smith <tsmith@chef.io>