Rename visudo_path -> visudo_binary and raise if the old one is used
They are entirely different concepts and users of the old one would see visudo never work. This feature went in 11/2017 and was never documented or added to the changelog, but internally we used it and there are probably a few users. We can just remove it sometime in the future, but I suspect very few will ever be impacted by this.

Signed-off-by: Tim Smith <tsmith@chef.io>