Remove path helpers from window_pagefile
I backported locate_sysnative_cmd initially, but after chatting with Stuart this isn't something we need anymore. This removes it and instead uses wmic.exe directly, which in testing works just fine. This also removes the windows friendly path helper which is really just a gsub on the slashes. Instead I changed the coerce to get us what we need in terms of path formats. I copied this to a Windows 2016 box and it works fine with both formats of paths. I'm making this same change to the windows cookbook.

Signed-off-by: Tim Smith <tsmith@chef.io>