Fix array parsing in windows_feature_dism / windows_feature_powershell
This fixes how we parse out the arrays and adds testing to make sure it's doing what we want.

Plus it properly continues on when the user has removed all local feature but specified an external source of those via the registry. That was a reported issue that came in on the cookbook after the rewrite.

Lastly this also fixes a bad method call to the dism method that came over when diffing from the cookbook.

Signed-off-by: Tim Smith <tsmith@chef.io>