add delegator for property_is_set? to providers
This was dropped in Chef 14.0.190 by chef/chef#6952 as a side effect
of stripping out all the Chef::Resource automagic delegators.

People were using this one inside of provider code, and it seems fine
to continue to use it.

Moving it to inject it directly into the Provider class instead of
just custom resource action_classes to make our API more consistent.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>