fix Chef-14 chef_fs/chef-zero perf regression
Fixes a perf bug introduced by #6471 where we were taking O(n^2)
with the number of cookbook versions in the cookbook synch phase.

This is a minimum viable fix that changes the algorithm so that
it does the old, fast method of looking up the CBV first.  Only
if it doesn't find it does it then go searching for cookbooks
that are in a differently named directory but which match by
the name in the metadata.

There may be edge conditions here if people mix-and-match and
have cookbooks which have both named-directory and
name-in-metadata versions.  Please try not to do that.  Don't
cross the streams.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>