add the resources() dsl method to providers
Custom resource actions were picking this up via delegation to
the wrapping resource, and that wiring was removed in 14, so that
API then got dropped accidentally.  This should fix that back
up.  It also consistently injects the resource APIs into resources
and providers both now (and consistently across core resources
and custom resources).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>