Run all tests in CI
Previously, we ran only integration tests. But many of our unit tests
are constrained by rootyness and platform, and this means the first time
they get run is during the ChefDK verification stage, which uses the
released Chef build. So we have no way of fixing test failures short of
a new Chef release.

Signed-off-by: Thom May <thom@chef.io>