Use some unique task names for windows_task functional tests
The 2008r2 tester has been failing with an error like:

```
  1) Chef::Resource::WindowsTask action :enable enables the disabled task
     Failure/Error: expect(current_resource.task.status).to eq("not scheduled")

       expected: "not scheduled"
            got: "running"

       (compared using ==)
     # ./spec/functional/resource/windows_task_spec.rb:1481:in `block (3 levels) in <top (required)>'
```

This has only been happening on the enable and disable tests and appears to be a timing issue. This change gives thoses tests unique names and changes the default name to not be 'chef-client' which is more likely to collide with another test.

This change has passed chef-test on jenkins in an ad-hoc test.

Signed-off-by: Bryan McLellan <btm@loftninjas.org>