Deprecated the Chef::Provider::Package::Freebsd::Pkg provider
FreeBSD 10 and later have pkgng and Chef will use that instead. This Oct even FreeBSD 10 goes fully EOL so we'll be supporting multiple EOL versions back come April. We should speed up this resource and simplify things by removing the unnecessary provider and the logic that decides which one to use.

Signed-off-by: Tim Smith <tsmith@chef.io>