Fixing #7347
Red
===
The pre-existing assertions around this functionality can probably be deleted after these tests are made green.

Green
=====
Wrote production code to pass tests, updated mocks on related tests

Added a new requirement for the checkout action to verify the target directory doesn't exist or is empty. Then in the `action_sync` method, changed the direct call to `action_checkout` to `run_action :checkout` so it goes through all the validation steps for that action. Also needed to update a couple of test's mocks to work with this new workflow. I'll refactor a bit in the next commit.

Refactor
=======
Removing duplication, updating old code

- There was a lot of duplication in the spec that I moved into RSpec `let` values.
- Replaced duplicate call to `::File.exist?(::File.join(cwd, ".git"))` with `existing_git_clone?`
- Replaced direct call to `action_checkout` in `action_export` with the more appropriate `run_action :checkout` to go through the proper workflows.
- Removed now unreachable code branch in `action_checkout`

Signed-off-by: Mike Brown <mikbro@microsoft.com>