Don't crash on non-existent credentials profile
If a credentials file exists, but doesn't contain the profile in
CHEF_PROFILE (or a default profile is CHEF_PROFILE is unset and there is
no context file), then an "Unable to parse Credentials file" error is
shown.

This change will print a better error message if the profile isn't found.
Additionally, in the case where there isn't a default profile, we just
fall back to normal credential loading without using anything in the
credentials file (similarly to if the credentials file didn't exist at
all).

Signed-off-by: Mark Harrison <mark@mivok.net>