Remove awesome customers testing and update kitchen configs
Awesome customers cookbook testing isn't even correct from learn chef anymore and it's not coming back in this form. There's no reason to keep carrying it around.

This change also deps on kitchen 1.21 and uses the new name for the configs. Much easier to find if you're browsing the source in Finder/Explorer.

Signed-off-by: Tim Smith <tsmith@chef.io>