Remove the existing acceptance testing framework
We haven't used this framework for 2 major releases of Chef and as time goes on it's more and more likely we'll build something very different for testing.

We 100% do need acceptance testing, but between dokken testing in Travis  and the potential for more complex integration testing with build-kite I think we should do the proper cleanup and remove this code now.

Signed-off-by: Tim Smith <tsmith@chef.io>