Mildly speed up knife by moving deps into dep blocks
We can lazily load the requirements for knife plugins, but we don't do that for all our plugins. We should. Is this a game changer? Nope. It does make it 4% faster though, which makes it worth doing.

Side note: We have some dupe requires in here and we also requires mixlib-shellout and mixin-shellout when we only used mixin-shellout.

Signed-off-by: Tim Smith <tsmith@chef.io>