fix el6 32-bit tests
We've been relying on the yum libraries to throw away the version
information that we supply here and return the version that is
installed that matches the name + arch.

On el6 32-bit we've hit a bug in the yum libraries where it appears
to not do this.  The simple solution here is to just throw away the
version information ourselves.

I think at one point early in the development of the new yum provider
that there was more of a point to supplying the version here, but as
it evolved that went away, but this argument was still kept around
vestigially.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>