Utilize repo discovery in Chef::ChefFS::Knife
When using a newly generated chef-repo + `knife configure` the Knife
commands that are based off of `Chef::ChefFS::Knife` currently fail
because the `Chef::Config[:chef_repo_path]` and associated sub-paths
default to be based off of the `Chef::Config.cache_path` which will be
`~/.chef`. This commit checks for this case and instead utilizes `discover_chef_repo`
to get a more accurate default chef_repo_path based on where the
knife command is being run from.

Signed-off-by: Keith Thompson <me@keiththomps.com>