Honor chef_server_url for sandbox uploads
Chef uploads cookbooks transactionally by generating a unique sandbox
URL to upload all the associated files to using a "POST
/organizations/<organization>/sandboxes" operation.

This POST operation returns the URL of the sandbox that should then be
used to PUT the cookbook's files to. This URL is pre-resolved, based on
the configuration of the chef server, which can be different from what
is configured in the client.

In special cases of network configuration, this discrepancy forces the
client to use a URL with a non-resolvable or unreachable host for
uploading cookbooks, rather than the scheme/host/port configured with
chef_server_url.

This change parses the sandbox URL returned by the server, and rewrites
it honor the chef_server_url configuration.

Semi-Related to: https://github.com/chef/chef-server/issues/50

Signed-off-by: Scott Bowers <sbbowers@gmail.com>