Throw better error on invalid resources actions
Right now if a user enters an incorrect action they get a message like
this that doesn't print the actions as being symbols. That's super
confusing to new users and results in random cookbook bugs from time to
time.

[2018-10-07T13:26:39-07:00] FATAL: Chef::Exceptions::ValidationFailed: Option action must be equal to one of: nothing, create, remove, modify, manage, lock, unlock!  You passed :whatever.

With this change we properly print out symbols as symbols so it's clear
what the available values are.

[2018-10-07T13:24:09-07:00] FATAL: Chef::Exceptions::ValidationFailed: Option action must be equal to one of: :nothing, :create, :remove, :modify, :manage, :lock, :unlock!  You passed :whatever.

It's a small thing, but I think this will have a pretty big impact on
first time users. This one tripped me up when I first started.

Signed-off-by: Tim Smith <tsmith@chef.io>