Node Attributes: Build ImmutableMash properly in deep_merge!
closes #7738

The root cause here is that there's a difference between doing

Chef::Node::VividMash.new( <another VividMash> )

And:

Chef::Node::ImmutableMash.new( <a VividMash> )

The former short circuits and does no work in convert_value.  The
latter will not short circuit and does the proper work to dup
and convert the value argument.

Since we build an ImmutableMash now, we do not need to wrap it
with the extra immutablize() call.

This should be perf neutral or very, very slightly better perf.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>