Add skip_docs and default_description to resource properties
skip_docs is used to specify a property that we don't want to include in the documentation. We have several of these.
default_description is used to describe the default in a way we'd want to put on the docs site. This is particularly useful for describing all our lazy / computed values.

I've used both of these in resources to show how I think we'll use them.

Signed-off-by: Tim Smith <tsmith@chef.io>