windows_share: Fix idempotency by removing the "everyone" access
This resource uses powershell under the hood and calls new-smbshare,
which defaults to adding read only access to the everyone group. With
this change when we create the share we'll remove that permission. Once
that's done we'll go about adding our desired permissions. This only
runs once so the overhead is pretty low and fixes idempotency.

Signed-off-by: Tim Smith <tsmith@chef.io>