ignore failures doing apt-get update
when the apt-get update fails in the before(:all) block it has the effect of
failing the block.  when the block fails the rspec will retry it again
on the next test, even though it is marked :all.

what we're seeing is that we're getting rate limited by the upstream apt
source.  when that happens, the before(:all) becomes a before(:each)
which causes it to hammer on apt-get update, which cannot help with the
rate limiting.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>