Disable s3 omnibus cache
This has caused a ton of random issues where builders start pulling in
cache from other branches. Every time we get in this state we just
rebuild without the cache and all is fine again. It seems like we should
just disable it at this point since we run builds on 2 branches at the
same time now.

Signed-off-by: Tim Smith <tsmith@chef.io>