Remove support for unsupported opensuse < 42 from group provider
We had a few places where we checked for legacy opensuse platforms. Chef 15 doesn't support this platform and you won't be able to find binaries via omnitruck. We might as well yank out a bit of provider logic.

Also for some reason usermod has logic in a method for suse, but the provider didn't support that.

Signed-off-by: Tim Smith <tsmith@chef.io>