Update knife bootstrap template to use up to date omnitruck URL
omnitruck-direct.chef.io is a legacy URL that now points to the same
place as omnitruck.chef.io, and was originally in place to support
clients with older SSL libraries. However, the backward compatible
behavior of omnitruck-direct hasn't been the case for a while now. This
change just corrects the URL in the bootstrap template also.

Signed-off-by: Mark Harrison <mark@mivok.net>