Remove travis gem dep and bump versions of multiple components
Bring in many newly released gems that loosen mixlib gem deps. This allows us to bring in the new major releases of these gems. The major version bumps were done to allow for the dropping of Ruby 2.1/2.2 so this is not a breaking change for Chef 14. It does get us some bug fixes we've made since the major bumps though.

Signed-off-by: Tim Smith <tsmith@chef.io>