Allow empty strings in -o to result in empty override run list
Previously this was not possible.  Now you can do this:

```
chef-client -o "" ./foo.rb
```

And it will only run foo.rb and not any of the recipes in your
run_list at all.  Node will not be saved at the end.

Useful to one-shot something like keyrotation (which still needs
keys and needs to talk to the chef server API, so this is a
different use case from simple chef-apply).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>