Remove a functional test for the reboot_pending DSL
The 'Component Based Servicing' registry key is limited by permissions
to Trusted Installer, so we can't easily create/delete it. We do cover
this key in the unit tests, and the other functional tests cover the
code paths already, so we can remove this rather than leaving it
'pending' forever.

Closes #3085

Signed-off-by: Bryan McLellan <btm@loftninjas.org>