Add train from git
This is temporary. Once inspec 4 gem is released,
correct deps will get pulled in, and we'll be
moving chef-core to also use train-core instead of train.

In the interim in order
to get appbundler builds passing for the bootstrap branch
with a train dep from git, this needs to be merged to master. This will
ensure the train dep is included in the package used to seed
appbundler-update.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>