Remove chef-* binstubs from chef gem
This is for the relicensing effort.

Note that this PR leaves the knife and ohai binstubs still in the
gem(s) while that discussion is still ongoing.

This PR extracts the binstubs to chef-binstubs gem which is not intended
to be redistributed outside of the omnibus chef binaries.

A blank gem for chef-binstubs has been pushed to rubygems.org for
obvious security concerns and so that bundler/rubygems can resolve
the gem

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>