Check for directories on Windows before creating
Under Chef 14 bootstrap, creation of directories
that already existed in Windows would log an error,
but not raise an error.

With the Chef 15 bootstrap change to run the bootstrap
script under powershell, any error in the script now causes
an error to be raised up. Even though the bootstrap complets
successfully - the script execution does not halt on error -
it presents a confusing error to the human that makes it look like
bootstrap failed.

Fixes CHEF-8469

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>