Use new Net:SSH host key verify values
Fixes #8482 by prompting to continue if the host key is not recognized.

This attempts to tell Net:SSH to accept and write the key but it is only
temporarily accepting.

This changes the parameters of --ssh-verify-host-key (which is new) to
take the Net::SSH values, which get passed through by train. This allows
the user more options than only verifying or not.

Signed-off-by: Bryan McLellan <btm@loftninjas.org>