Add a warning when session-timeout is too short
We converted --winrm-session-timeout MINUTES
to --session-timeout SECONDS

If someone is providing a short session timeout, it's possible
that they're doing so based on the flag accepting minutes.  We'll warn
them and suggest an alternative so that if things go wrong,
they'll know where to start looking.

We can remove this once we remove the deprecation for
winrm-session-timeout

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>