Bootstrap updates
This includes a few fixes:

1. properly handle the default value for session_timeout in cases where
merge_config is not called (plugins) to populate defaults.
2. verify that we're using the ssh protoocl before referencing ssh
constants.
3. capture and compare the higher-level Train::Error for the check
to see if we've gotten a missing fingerprint error, instead of using
Train::Transports::SSHFailed which may not be loaded when exception
is being evaluated.
4. Add tests for missing fingerprint behavior
5. do not save session_timeout to Chef::Config[:knife], it was not
referenced as knife config in any location and we're trying to move away
from pushing CLI values into knife config.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>