(GH-8580) Add ability to toggle Chocolatey feature
Add the ability to enable and disable a named Chocolatey feature by
creating a new Chef resource named chocolatey_feature.  This accepts a
single property either from feature_name or directly from name attribute
then, based on the defined action, will either enable or disable that
feature.  Resource inspects the current state of the chocolatey.config
in order to decide whether a change is required or not.

Signed-off-by: Gary Ewan Park <gep13@gep13.co.uk>