Pass yum_package options to underlying python helper
When working with custom repo that are disabled by default, we need to
enable them on-demand. The python helper handles that properly but the
yum package provider does not pass user's options in all cases.

Most of the time there is no issue because in normal cases you should
not need to enable any repository to work with "installed" packages.

However in some rare cases, you may have multiple installed versions of
a given package. In that situation, python libraries used by the helper
will do some extra work that may (will) requires to enable the repo.
See call to _compare_providers in http://yum.baseurl.org/download/docs/yum-api/3.2.27/yum-pysrc.html#YumBase._bestPackageFromList

Not passing the options in that case, will result in an unclear error:
  > ' * No candidate version available for "package" '
And often will lock/corrupt the rpm db.

Signed-off-by: Baptiste Courtois <b.courtois@criteo.com>