WIP: the beginning of a hab package sanity check
Modeled after the Habitat core-plans testing standard[1], these are
tests for sanity checking the viability of a chef-client habitat
package. The scope of viability is to consider a package promotion from
unstable to current. I would like to emphasize a line from there on
scope of testing:

> if a failing test would not be fixed via a change to [the habitat
> plan], the test does not belong in this [set of package tests].

In other words, unit and functional testing of the code should occur
before build. Functional/integration testing could be added to these
tests later or be performed in a similar fashion in a different set of
tests for deciding whether to promote from current to stable.

[1] https://github.com/habitat-sh/core-plans/blob/master/docs/dev/policy_documents/standards-for-tests.md

Signed-off-by: Robb Kidd <robb@thekidds.org>