consolidate some options in the base class
most of these are options that apply to either chef-client -z or
to chef-solo without legacy mode, so they are justifiably common
options (that are handled by the trampolining into the Client
code from Solo right now.

the --legacy-mode switch is solo-specific (although we should wire it
up for chef-client eventually once all the code is merged).

old-reporting and skip-cookbook-sync only make sense for actual
chef-client runs so are marked as such in the help text

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>