Property: Reorder comparison with NOT_PASSED.
The original comparison `value == NOT_PASSED` ends up calling `.<=>` on
value under the hood. This relies on the implementation of `<=>`/`cmp`
not being broken in that class.

Unfortunately, some classes like OpenSSL::X509::Name have broken
implementations of `<=>` that raise TypeError instead of returning nil
for mismatched types. https://github.com/ruby/openssl/pull/264

As a result, currently, if you attempt to set a resource property to an
OpenSSL::X509::Name object, the result is a TypeError.

With this change, we use the implementation of comparison from the
NOT_PASSED object instead, which we know will work no matter what.

Signed-off-by: Andy Brody <git@abrody.com>