Consistently use NOT_PASSED over alternatives
We have four different versions of this now, and this converts to
just using the one convention in `lib/chef/constants.rb`

Interestingly `Chef::NIL_ARGUMENT` is used nowhere in the codebase, that
probably got refactored out in 12.5.1.

Node objects still use `NIL` but that is a different kind of `Object.new`
which really is semantically private to the implementation of
attributes.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>