Always use system rubygem to determine nodocument flag.
(This could only happen on chef `14.x` at this moment)

Currently if the user doesn't specify custom rubygem binary path,
chef uses `Gem::VERSION` to determine if the rubygem version is
compatiable with old nodocument flags like `--no-rdoc`.

https://github.com/chef/chef/blob/7558c414030d082194d0d2b5f74279ce151517e9/lib/chef/provider/package/rubygems.rb#L601-L611

https://github.com/chef/chef/blob/7558c414030d082194d0d2b5f74279ce151517e9/lib/chef/provider/package/rubygems.rb#L272-L274

https://github.com/chef/chef/blob/7558c414030d082194d0d2b5f74279ce151517e9/lib/chef/provider/package/rubygems.rb#L303-L305

This doesn't work because the `Gem::VERSION` always points to chef's
embedded rubygem version. But the binary we use to install gems is
the system's rubygem, which might be different from chef's embedded
version even though the user doesn't specify custom binary path.

So when a user upgrades his Ruby version to 2.6 (rubygem 3.0.x) with
Chef 14.x (rubygem 2.7.x), an error will occur:

```
 STDOUT:
   STDERR: ERROR:  While executing gem ... (OptionParser::InvalidOption)
       invalid option: --no-rdoc
```

This is because `Gem::VERSION` would make chef think that it's ok to use
`--no-rdoc` while it's not. The solution is to always perform the check with
system rubygem's version.

Signed-off-by: st0012 <stan001212@gmail.com>