Fix node[:cookbooks] attribute
closes #8817

Note that if any calling code winds up seeing this error message:

```
NoMethodError: undefined method `set_cookbook_attribute' for nil:NilClass
```

That means that the cookbook_collection was set before the node was
set on the run_context.  That wouldn't be a bug in core chef, that must
be fixed in the caller to reverse the order of operations.

Since I only made the positional arguments to the run_context constructor
optional in Chef-15.0 though I don't expect this breaks any existing code
written in the past month or two, but if anything crops up in the future,
consider this a definitive statement that the caller must reverse the
order of their operations and this error being thrown is a feature not
a bug to be fixed.

(The fact that we silently aborted rather than threw a NoMethodError on
NilClass meant that we shipped this defect -- sometimes defensive
programming can be overly defensive and swallow real errors).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>