remove app_server_support spec file
Convert to using TinyServer and remove the duplicated functionality
with the "app_server_support" helper.

The fact that this change hits the knife raw / knife redirection specs
that have so many timeout issues is very likely not a coincidence.

I would bet a small, but not entirely insignificant amount of money that
this makes those timeout issues go away.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>