update syntax of `update-rc.d` commands in enable & disable actions
Previous implementation was built for an outdated version of
`update-rc.d` and it used commands that no longer work.

The `stop` and `start` commands were removed in 2.88.dsf-42 [1]:

```
  * sysv-rc:
    - update-rc.d no longer supports non-dependency-based boot.
      + Remove non-insserv codepaths.
      + Warn if the start or stop actions are used.
      + Skip runlevel mismatch warnings if default action is used
        (no arguments to validate).
      + Update manual page to remove start and stop actions, plus
        manual setting of boot sequence ordering; note that start
        and stop are no longer supported.  Closes: #606505.
```

and has been since ported to all currently supported Debian
distributions (including Jessie, released in 2015).

Both `start` and `stop` commands now effectively call `defaults`.

As a result of this `action_disable` did not disable a service, it would
enable it instead. Chef would execute the following commands on disable:

```
/usr/sbin/update-rc.d -f <service_name> remove
/usr/sbin/update-rc.d -f <service_name> stop 80 2 3 4 5 .
```

But `update-rc.d` would effectively run:

```
/usr/sbin/update-rc.d -f <service_name> remove
/usr/sbin/update-rc.d <service_name> defaults
```

And the service gets enabled instead.

With changes in this commit Chef will effectively run the following
commands:

```
/usr/sbin/update-rc.d -f <service_name> remove
/usr/sbin/update-rc.d <service_name> defaults
/usr/sbin/update-rc.d <service_name> disable
```

The service is now disabled, as expected.

Additionally `priority` support has been dropped entirely from code as
it's not supported by `update-rc.d` either.

It's potentially a breaking change on a very outdated distributions and
it could be worked around with additional `if/else` branches, but given
all non-EOLed distros have up-to-date version of `update-rc.d` available
it didn't feel worth the complexity.

[1] https://launchpad.net/debian/+source/sysvinit/2.88dsf-42

Signed-off-by: Rob Ulejczyk <rulejczyk@gmail.com>