Modify #8901 to not use arrays with mixlib-shellout
Chef-14 has to work with mixlib-shellout 2.x which does not support the array form of shelling out. Use a string instead. I also updated the gitignore to avoid some junk coming in and swapped the spec's bundler require to be the full bundler gem since it needs a giant pile of stuff in Bundler and bundler doesn't have proper requires in the DSL class to make that happen.

Signed-off-by: Tim Smith <tsmith@chef.io>