anonymize published cookbook contents
The tar commands default to storing the user and group IDs--and the
user/group names associated with them--of the directories and files
being archived. This change updates the BSD and GNU forms for the tar
command to set the owner and group to 0 and to not store the user/group
name mapping in the resultant archive.

There is also has a benefit of avoiding bugs in tar readers that do not
behave well when large uid/gid values are stored in tar's extended
metadata headers. While this won't fix cookbooks that have already been
published, it will minimize the bug's appearance in clients with older
dependency resolvers as fewer latest cookbook versions will include
large IDs.

Signed-off-by: Robb Kidd <robb@thekidds.org>