depend on bash because we shell out to it *a lot*
bash isn't on the package's PATH when executed via "hab pkg exec." While
the system's bash will be on the PATH when chef-client is run as a
Hab-supervised service, the functional specs can't be executed as a
hab-sup'd process. So, for the time being, depend on bash so that it's
around when the functional specs run.

Signed-off-by: Robb Kidd <robb@thekidds.org>