add a new pipeline for verifying hab plans (#9019)
As noted in the pipeline's description, this is intended to be a
temporary pipeline to verify the habitat plans. It will soon perform a
local build, test it, then throw it away.  Its steps should eventually
move into the "verify" pipeline (if fast enough) or this pipeline should
be removed in favor of learning about packaging problems after merge.

definition is overridden to get specific about where the steps are
defined.

trigger is overridden so that the usual `verify/` trigger of PRs
opened/updated is turned off. This pipeline will only be triggered by
API call (and so, also through the BuildKite UI once the pipeline
exists).

Signed-off-by: Robb Kidd <robb@thekidds.org>