move minimum hab check/install to a reusable script
And use it in both the verify/habitat and habitat/test pipelines.

verify/habitat needs hab >= 0.85 to both build with environment
variables that don't include the studio path[1] and to binlink the runtime
with stubs that include the package and system environment.[2]

habitat/test needs hab >= 0.85 to binlink the runtime with stubs that
include the package and system environment.

[1] https://github.com/habitat-sh/habitat/pull/6850
[2] https://github.com/habitat-sh/habitat/pull/6826

Signed-off-by: Robb Kidd <robb@thekidds.org>