Add new chef_sleep resource
We highly recommend you never use this, but there's times when you 100%
need to have the client sleep for a period of time after some work.
Example: windows_service will start a service and the OS will say that
it's "started" when it's not. Before doing additional work that relies
on the service being started you may need to sleep. This would be a good
time to notify on the restart or start action.

Signed-off-by: Tim Smith <tsmith@chef.io>