Fix mac_user breakage and logging
We need more defensive coding around the absense of the ShadowData.

Unfortunately I wasn't able to fix the tests on my mac where this
happens, and don't know how to introduce tests that would fail on
buildkite, so we have no coverage for this.  Testing was done
manually on the CLI.

Also removed some useless converge_by statements that were
effectively double-logging.  The methods that the converge_by's
were in are already wrapped by converge_by's in the superclass
so the resource is already "dirty" in the sense that it knows
that it is being updated, and the log message that e.g. the
user is being removed is already being printed.  Some of the
sub-action logging with sub-converge_by's here is still good
since it shows the details of how the user is being added or
removed or modified which makes sense (still sort of abusing
converge_by to get only logging, but that doesn't matter because
we're in a block where we MUST be updating the resource anyway).

This commit also re-adds testing of unit+integration testing to the
omnibus pipeline.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>