Use the right class in knife supermarket installl
A user on discourse is seeing errors related to this old class. I'm not able to reproduce that failure, but we shouldn't be calling the deprecated class here. Changing this avoids a bogus deprecation warning that tells the user to stop using the legacy command, when they're actually using the new command.

Signed-off-by: Tim Smith <tsmith@chef.io>