Use /etc/chef for bootstrapping instead of CONF_DIR
CONF_DIR is resolved based on the host system running chef-client/knife.  When we
use it in bootstrap template/context, we need a value based on the
target system that is being bootstrapped so that the paths are correct
(so that we don't try to create C:\chef on a linux system).

Fixes #9224

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>