Swap the methods and the aliases in the chef-utils platforms
X_platform? is the name we want to eventually use so make that the real
method name as make the alias the name we want people to stop using.
That way we document the right thing in yard and can skip all the
aliases when we generate the vscode snippets.

Signed-off-by: Tim Smith <tsmith@chef.io>