Make sure we don't break the world in RHEL 9
I was thinking about how to accomplish this a bit. At first I figured we should use the which helper and maybe just make a dnf? question helper, but we really care about RHEL 8 here and not RHEL 7 if it has DNF. Also since this only works on Redhat we don't have to worry about all the other platforms that also have DNF since they don't apply to this resource.

Signed-off-by: Tim Smith <tsmith@chef.io>