Fix file descriptor leak in our tests
Mac has a default max fd of 256 which this can pretty much blow out by
leaking PIPE objects which last until the garbage collector destroys
the instances, this forces the pipes to be closed when the stop function
is called here.  Cuts down on over 200 leaking file descriptors from
this test which leak into the subsequent tests.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>