CONTROVERSIAL: remove "with a chef repo" shared context.
It's provided by cheffish, but with one difference (force).

Unfortunately, that force can't be pushed to cheffish as-is because it
relies on ChefUtils.windows?. This could be changed to something
else... A global for all I care.

But, we might not want to push this commit yet. I just can't stand the
noise in the test output as I try to clean this stuff up.

Signed-off-by: Ryan Davis <zenspider@chef.io>