Fix functional tests on Windows 10 by matching less
Windows 10 1909 is a service pack, but without being called a service pack.

1909 is a very small update over 1903, and in fact most of the binaries are not rebuilt.

The version for 1909 is 10.0.18363, but the version in many files is still 10.0.18362, which was the version for 1903.

This caused this functional test to fail on 1909 because 10.0.18362 != 10.0.18363. Nowhere else are we concerned with the file versions matching the OS version, so the fix is to match less of the version. By matching only '10.0', we should not see this again.

Signed-off-by: Bryan McLellan <btm@loftninjas.org>