Support non-Linux hosts in chef_client_cron
The cron_d resource only works on systems that support and /etc/cron.d directory, which is Linux as far as I can tell. We want this resource to work on all the other *nix platforms so we need to fall back to the legacy cron resource for those hosts.

Signed-off-by: Tim Smith <tsmith@chef.io>