Make sure all the non-multipackage packge resources can't take arrays
Right now these resources inherit package_name, versions, and options from the package resource which allows arrays. The problem is the actual providers for these package resources don't know what to do with arrays and they explode. They explode in ways that are really unfriendly to users. With this change they'll get a nice input validation error and they'll have a lot more of an idea of how to go about resolving the problem. This also means that our automatically generated documentation for the package resources will be correct since right now it documents everything as accepting an array.

Signed-off-by: Tim Smith <tsmith@chef.io>